Add mechansism to use different classloaders
Openfire plugins get typically loaded in a dedicated classloader (or re-use the classloader of a parent plugin, when one is defined).
The OFMeet plugin will consist of serveral 'modules' that are third party (mostly Jitsi) projects. As these projects were designed to
run in a dedicated JVM, placing them together in one JVM leads to issues. One of these issues is that the dependencies of each of the
Jitsi projects can conflict with dependencies of another Jitsi project.

To allow these projects to all be used in one Openfire plugin, this commit introduces a mechanism where each Jitsi project gets loaded
by its own class loader. This class loader is a custom implementation (that borrows from Openfire's PluginClassLoader). This in theory
would allow all Jitsi projects to co-exist within one Openfire plugin.

As a proof-of-concept, this PR applies the mechanism to the JVB project that was already part of the ofmeet plugin. At a later stage,
additional projects will be added.