Reworked module packaging
For some reason, the lib directory for each module still included libraries that were intended
for other modules only. This commit changes the approach to building the assembly. Now, every
module builds its own module-assembly, which is attached as a maven artifact, and used by the
ofmeet assembly to create the Openfire plugin.