Guard against running with OFFocus
This version of OFMeet no longer requires the OFFocus plugin. The presence of OFFocus plugin in the system
will likely negatively affect the functionality of the current version of OFMeet. This commit adds a warning,
when both plugins are installed, preventing the OFMeet plugin from loading if it does, and instructing the
administrator to remove OFFocus.