Use upstream lib-jitsi-meet again.
This commit undoes the hack that was introduced to fix https://github.com/jitsi/libjitsi/pull/461

In earlier commits, the jitsi components have been updated to include a version of lib-jitsi-meet
that includes the fix (and actually has more improvements that we'd like to pull in). We should no
longer pin lib-jitsi-meet.